fix: missing sessionId on watchlist api call
Even though the API doc marks sessionId as not required, it turns out (and makes sense) that it is necessary to pass it to verify the account holder. It is weird however that you can make the call without passing sessionId. In that case, the call returns the results for the owner of the API key account!!! (?)