fix: replace jellyseerr bottom sheet with screen
There's an issue with bottom sheet behaviour, and the long press text actions "Copy", "Cut", "Paste", "Select all" etc.. are not visible on text field on top of bottom sheet. Therefore we decided to remove bottom sheet and replace it with a screen. This improves the users' experience as it'll be easier for them to copy-paste their credentials.

Google issue: https://issuetracker.google.com/issues/307329984