feat: make database insertion run on parallel scopes
For longer credits responses, we optimized the process by running database insertion in parallel with API emission. We don't need to wait for the database to finish before updating the UI. Therefore, we created a new CoroutineScope to launch the insertion process. Additionally, each DAO call runs in a separate scope, reducing the total process time from 15,000ms to just 4ms!